Move MacRoman encoding helpers to core crate
Allows them to be used in the BlueSCSI Toolbox implementation, so that
(some) special characters in transferred file names can be preserved.
