+++ core/Cargo.toml
@@ -58,6 +58,7 @@ hex = "0.4.3"
 arpfloat = { version = "0.1.11", default-features = false, features = ["std"] }
 serde-big-array = "0.5.1"
 typetag = "0.2.20"
+unicode-normalization = "0.1.24"
 postcard = { version = "1.1.3", features = ["use-std"], optional = true }
 binrw = { version = "0.15.0", optional = true }
 zstd = { version = "0.13.3", optional = true }
